requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.2.0
comm==0.1.2
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
filelock==3.12.2
fonttools==4.40.0
future==0.18.3
idna==3.4
ipykernel==6.21.3
ipython==8.11.0
iso8601==1.1.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.1.1
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.10.3
Pygments==2.14.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==25.0.0
requests==2.31.0
SciPy==1.10.1
serial==0.0.97
six==1.16.0
stack-data==0.6.2
sympy==1.12
torch==2.0.1
torchvision==0.15.2
tornado==6.2
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.3
wcwidth==0.2.6
